+++ src/xenia/core/memory.h
@@ -29,10 +29,10 @@ xe_memory_ref xe_memory_retain(xe_memory_ref memory);
 void xe_memory_release(xe_memory_ref memory);
 
 size_t xe_memory_get_length(xe_memory_ref memory);
-uint8_t *xe_memory_addr(xe_memory_ref memory, uint32_t guest_addr);
+uint8_t *xe_memory_addr(xe_memory_ref memory, size_t guest_addr);
 
-uint32_t xe_memory_search_aligned(xe_memory_ref memory, uint32_t start,
-                                  uint32_t end, const uint32_t *values,
+uint32_t xe_memory_search_aligned(xe_memory_ref memory, size_t start,
+                                  size_t end, const uint32_t *values,
                                   const size_t value_count);
 
 // These methods slice off memory from the virtual address space.
